% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comat_to_df.R
\name{comat_to_df}
\alias{comat_to_df}
\title{Create a data.frame from a contingency table}
\usage{
comat_to_df(comat, weight = FALSE, remove_absent_species = TRUE)
}
\arguments{
\item{comat}{a contingency matrix with sites as rows and species as columns}

\item{weight}{a boolean indicating if the value are presence (i.e. equal 1) or abundance data}

\item{remove_absent_species}{a boolean determining whether absent
species from the co-occurence matrix have to be removed from the output.}
}
\value{
A data.frame  where each row represents the presence of a species (second column)
in a site (first column) and an optional third column indicating the abundance
of the species in this site
}
\description{
This function creates a two- or three-columns data.frame where
each row represents the presence of a species (second column)
in a site (first column) and an optional third column indicating the abundance
of the species in this site (if weight=TRUE) from a co-occurence matrix
(sites as rows and species as columns).
}
\examples{
comat=matrix(sample(1000,50),5,10)
rownames(comat)=paste0("Site",1:5)
colnames(comat)=paste0("Species",1:10)

df=comat_to_df(comat,weight=TRUE)
}
\author{
Pierre Denelle \email{pierre.denelle@gmail.com},
Maxime Lenormand \email{maxime.lenormand@inrae.fr}
and Boris Leroy \email{leroy.boris@gmail.com}
}
