% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_comat.R
\name{df_to_comat}
\alias{df_to_comat}
\title{Create a co-occurence matrix from a data.frame}
\usage{
df_to_comat(df, weight = FALSE)
}
\arguments{
\item{df}{a two- or three-columns data.frame where each row represents the presence of
a species (second column) in a site (first column) and an optional third column indicating
the abundance of the species in this site}

\item{weight}{a boolean indicating if the abundance should be considered}
}
\value{
A data.frame  where each row represents the presence of a species (second column)
in a site (first column) and an optional third column indicating the abundance
of the species in this site
}
\description{
This function creates a co-occurence matrix with sites as rows and species as columns
from a two- or three-columns data.frame where
each row represents the presence of a species (second column)
in a site (first column) and an optional third column indicating the abundance
of the species in this site (if weight=TRUE).
}
\examples{
df <- data.frame(Site = c(rep("A", 2), rep("B", 3), rep("C", 2)),
Species = c("a", "b", "a", "c", "d", "b", "d"),
Weight = c(10, 100, 1, 20, 50, 10, 20))

comat=df_to_comat(df,weight=TRUE)
}
\author{
Pierre Denelle \email{pierre.denelle@gmail.com}
Maxime Lenormand \email{maxime.lenormand@inrae.fr}
Boris Leroy \email{leroy.boris@gmail.com}
}
