% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_contingency.R
\name{df_to_contingency}
\alias{df_to_contingency}
\title{Create a contingency table from a data.frame}
\usage{
df_to_contingency(df, weight = FALSE)
}
\arguments{
\item{df}{a two- or three-columns data.frame where
each row represents the interaction between two objects (site and species for example)
and an optional third column indicating the weight of the interaction}

\item{weight}{a boolean indicating if the abundance should be considered}
}
\value{
A contingency table with the first objects (first column of df) as rows and
the second objects (second column of df) as columns
}
\description{
This function creates a contingency table from a two- or three-columns data.frame where
each row represents the interaction between two objects (site and species for example)
and an optional third column indicating the weight of the interaction (if weight=TRUE).
}
\examples{
df <- data.frame(Site = c(rep("A", 2), rep("B", 3), rep("C", 2)),
Species = c("a", "b", "a", "c", "d", "b", "d"),
Weight = c(10, 100, 1, 20, 50, 10, 20))

comat=df_to_comat(df,weight=TRUE)
}
\author{
Pierre Denelle \email{pierre.denelle@gmail.com}
Maxime Lenormand \email{maxime.lenormand@inrae.fr}
Boris Leroy \email{leroy.boris@gmail.com}
}
