\name{run_oslom}
\alias{run_oslom}
\title{Run OSLOM algorithm on a projected network.}
\usage{
run_oslom(dat, n_runs = 1, t_param = 0.1, cp_param = 0.5, hr = 0,
saving_directory)
}
\arguments{
\item{dat}{a .tp object generated by OSLOM algorithm. Import in R with
readLines() function. This object can be generated with run_oslom() function.}

\item{n_runs}{Numeric parameter setting the number of runs for the first
hierarchical level. The default value is 10. The bigger, the slower the
function but the more accurate the results.}

\item{t_param}{Numeric parameter setting the probability value to consider
a module as significant. Must be comprised between 0 and 1.}

\item{cp_param}{Numeric parameter setting the resolution. Must be comprised
between 0 and 1.}

\item{hr}{Numeric parameter setting the number of hierarchical levels.
Must be a positive integer.}

\item{saving_directory}{Path where the OSLOM .tp file containing
         the bioregions identified will be saved.}

}
\description{
This function applies on a data.frame containing 3 columns. The first two columns
contain each pair of nodes of the network. The third column uses some metric to
quantify the similarity between each pair.
On this object, run_oslom() applies the OSLOM algorithm that identify bioregions
based on sites similarity. The code of the function is a compilation of the algorithm
provided on http://www.oslom.org.
It saves a .tp file with the identified bioregions in a chosen directory.
See http://www.oslom.org/code/ReadMe.pdf for more precisions.
}
\details{}
\section{Note}{}

\examples{
}
